# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Node API",
    "description": "Template API project for Node.js",
    "version": "1.0.0"
  },
  "basePath": "/api/node",
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/v1/persons": {
      "get": {
        "tags": ["person"],
        "operationId": "getPersons",
        "summary": "All the Persons",
        "description": "Return all the stored person-objects.\n\n_(Person.getPerson)_\n\n",
        "responses": {
          "200": {
            "description": "Persons",
            "schema": { "$ref": "#/definitions/PersonObject" }
          },
          "401": {
            "description": "This endpoint is secured - **credentials are missing or invalid**."
          }
        },
        "default": {
          "description": "Unexpected response - internal error"
        },
        "security": [{ "api_key": ["read"] }]
      }
    },
...
    "/v1/rooms": {
      "get": {
        "tags": ["room"],
        "operationId": "getRooms",
        "summary": "All the Rooms",
        "description": "Returns all the stored room-objects.\n\n_(Room.getRooms)_\n\n",
        "responses": {
          "200": {
            "description": "Rooms",
            "schema": { "$ref": "#/definitions/RoomObject" }
          },
          "401": {
            "description": "This endpoint is secured - **credentials are missing or invalid**."
          }
        },
        "default": {
          "description": "Unexpected response - internal error"
        },
        "security": [{ "api_key": ["read"] }]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
...
  "definitions": {
    "PersonObject": {
      "properties": {
        "id": { "type": "string", "example": "123" },
        "firstName": { "type": "string", "example": "John" },
        "lastName": { "type": "string", "example": "Doe" }
      }
    },
    "RoomObject": {
      "id": { "type": "string", "example": "123" },
      "type": "object",
      "required": ["name", "isBooked"],
      "properties": {
        "name": {
          "type": "string",
          "example": "Room A"
        },
        "isBooked": {
          "type": "boolean",
          "example": false
        },
        "relation": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "List of associated person IDs",
          "example": ["1", "2"]
        }
      }
    }
  }
}